requirements.txt
requests

django
djangorestframework
djangorestframework[authtoken]
django-cors-headers
